refactor(file structure): changed file structure to fit the standards
