+++ vite-project/src/config/NinjaApi.ts
@@ -1,0 +1,3 @@
+const API_NINJAS_API_KEY = "09CE2n85ib4gM6jB63HDOQ==5Xd52LR3dYw458MQ";
+
+export default API_NINJAS_API_KEY;
